+++ packages/gatsby/src/utils/worker/child.js
@@ -1,5 +1,8 @@
 // Note: this doesn't check for conflicts between module exports
+import { getFilePath } from "./page-data"
+import { renderHTML } from "./render-html"
+
 module.exports = {
-  ...require(`./page-data`),
-  ...require(`./render-html`),
+  getFilePath,
+  renderHTML,
 }