+++ packages/gatsby-plugin-preact/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@3.2.0...gatsby-plugin-preact@3.2.1) (2020-03-23)
+
+**Note:** Version bump only for package gatsby-plugin-preact
+
+# [3.2.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@3.1.28...gatsby-plugin-preact@3.2.0) (2020-03-20)
+
+### Features
+
+- **gatsby:** bump node min version to 10.13.0 ([#22400](https://github.com/gatsbyjs/gatsby/issues/22400)) ([83d681a](https://github.com/gatsbyjs/gatsby/commit/83d681a))
+
+## [3.1.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@3.1.27...gatsby-plugin-preact@3.1.28) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-preact
+
 ## [3.1.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@3.1.26...gatsby-plugin-preact@3.1.27) (2020-03-06)
 
 **Note:** Version bump only for package gatsby-plugin-preact